saving

+++ src/Windows_uSleep.h
@@ -1,6 +1,8 @@
 #ifndef WINDOWS_USLEEP_H
 #define WINDOWS_USLEEP_H
 
-extern void usleep(__int64 usec);
+
+
+extern void usleep(long long usec);
 
 #endif
